+++ front/src/pages/myPage/mypage.type.ts
@@ -1,6 +1,4 @@
 export interface propsType {
-  handleEnterMeeting: (_: React.MouseEvent<HTMLButtonElement>) => void;
   handleCreateMeeting: (_: React.MouseEvent<HTMLButtonElement>) => void;
-  meetingId?: string;
-  onChange?: () => {};
+  handleEnterMeeting?: (_: React.MouseEvent<HTMLButtonElement>) => void;
 }
